Move from in-file to external file
With this commit the hash of the translated file is no longer stored in
the translated file itself but in a file on the top level. That way we
are no longer required to modify the file itself. That prevents git-svn
from getting a hickup on rebase as suddenly the files are modified and
do not match the internal md5-hash any more